refactor(ui): split readiness hero presentation
